requirements.txt
# AI Capstone Project - Requirements
# Land Classification using Satellite Imagery with CNNs and Vision Transformers

# Core Deep Learning Frameworks
tensorflow==2.19
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# Scientific Computing
numpy==1.26
scipy==1.11.4

# Machine Learning
scikit-learn==1.7.0

# Data Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Progress Bars
tqdm==4.66.2

# Image Processing
Pillow==10.2.0

# HTTP Requests
httpx==0.27.0

# IBM Skills Network (for dataset download)
skillsnetwork

# Jupyter
jupyter==1.0.0
ipykernel==6.29.2
ipywidgets==8.1.2

# Utilities
python-dateutil==2.8.2
